+++ src/App.vue
@@ -148,6 +148,7 @@ export default {
         clearTimeout(this.refreshTimeout)
       }, 1000)
       this.setAgents(null)
+      this.setSortedAgents(null)
     },
     async connected(val) {
       if (val) {
@@ -250,12 +251,12 @@ export default {
       skip() {
         return (this.isCloud && !this.isAuthorized) || !this.connected
       },
-      pollInterval: 3000,
+      pollInterval: 1000,
       // Without this, server UI with no actual server shows results
       fetchPolicy: 'no-cache',
       update(data) {
         if (!data?.agent || this.isLoadingTenant) return null
-        this.setAgents(data.agent)
+        this.setSortedAgents(data.agent)
         return data.agent
       }
     })
@@ -395,7 +396,7 @@ export default {
     // window.addEventListener('focus', this.handleVisibilityChange, false)
   },
   methods: {
-    ...mapMutations('agent', ['setAgents']),
+    ...mapMutations('agent', ['setAgents', 'setSortedAgents']),
     ...mapActions('api', ['getApi', 'monitorConnection', 'setServerUrl']),
     ...mapActions('auth', ['authenticate', 'authorize']),
     ...mapActions('data', ['resetData']),