+++ Changelog.md
@@ -4,6 +4,7 @@
 
 ### Features and Improvements
 
+- Add agent screens - [#852](https://github.com/PrefectHQ/ui/pull/852)
 - Add the `mapped_children` route to the IAPI allow-list - [#891](https://github.com/PrefectHQ/ui/pull/891)
 - Improve expired auth and id token handling - [#896](https://github.com/PrefectHQ/ui/pull/896)